shuoz/requirements.txt
Django
gunicorn
gevent
django-markdownx
mysqlclient
